+++ src/ShopgateModule.php
@@ -7,7 +7,11 @@ namespace Shopgate\Shopware;
 use Doctrine\DBAL\Connection;
 use Shopgate\Shopware\Shopgate\Order\ShopgateOrderDefinition;
 use Shopgate\Shopware\System\Configuration\ConfigBridge;
-use Shopgate\Shopware\System\Db\PaymentMethodInstaller;
+use Shopgate\Shopware\System\Db\Installers\DeliveryTimeInstaller;
+use Shopgate\Shopware\System\Db\Installers\PaymentMethodInstaller;
+use Shopgate\Shopware\System\Db\Installers\RuleConditionInstaller;
+use Shopgate\Shopware\System\Db\Installers\RuleInstaller;
+use Shopgate\Shopware\System\Db\Installers\ShippingMethodInstaller;
 use Shopware\Core\Framework\Plugin;
 use Shopware\Core\Framework\Plugin\Context\InstallContext;
 use Shopware\Core\Framework\Plugin\Context\UninstallContext;
@@ -24,12 +28,17 @@ class ShopgateModule extends Plugin
             ConfigBridge::SYSTEM_CONFIG_PROD_EXPORT,
             [ConfigBridge::PROD_EXPORT_TYPE_SIMPLE, ConfigBridge::PROD_EXPORT_TYPE_VARIANT]
         );
+        (new RuleInstaller($this->container))->install($installContext);
+        (new RuleConditionInstaller($this->container))->install();
+        (new DeliveryTimeInstaller($this->container))->install($installContext);
+        (new ShippingMethodInstaller($this->container))->install($installContext);
         (new PaymentMethodInstaller($this->container))->install($installContext);
         parent::install($installContext);
     }
 
     public function uninstall(UninstallContext $uninstallContext): void
     {
+        (new ShippingMethodInstaller($this->container))->deactivate($uninstallContext->getContext());
         (new PaymentMethodInstaller($this->container))->deactivate($uninstallContext->getContext());
         parent::uninstall($uninstallContext);
 
@@ -46,12 +55,14 @@ class ShopgateModule extends Plugin
 
     public function activate(Plugin\Context\ActivateContext $activateContext): void
     {
+        (new ShippingMethodInstaller($this->container))->activate($activateContext);
         (new PaymentMethodInstaller($this->container))->activate($activateContext);
         parent::activate($activateContext);
     }
 
     public function deactivate(Plugin\Context\DeactivateContext $deactivateContext): void
     {
+        (new ShippingMethodInstaller($this->container))->deactivate($deactivateContext->getContext());
         (new PaymentMethodInstaller($this->container))->deactivate($deactivateContext->getContext());
         parent::deactivate($deactivateContext);
     }