+++ src/System/Db/Installers/DeliveryTimeInstaller.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Shopgate\Shopware\System\Db\Installers;
+
+use Shopgate\Shopware\System\Db\Shipping\GenericDeliveryTime;
+
+class DeliveryTimeInstaller extends EntityInstaller
+{
+    protected $entityInstallList = [GenericDeliveryTime::class];
+    protected $entityName = 'delivery_time';
+}