+++ src/System/Db/Installers/ShippingMethodInstaller.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Shopgate\Shopware\System\Db\Installers;
+
+use Shopgate\Shopware\System\Db\Shipping\GenericShippingMethod;
+
+class ShippingMethodInstaller extends EntityChannelInstaller
+{
+    use EntityActivateTrait;
+
+    protected $entityInstallList = [GenericShippingMethod::class];
+    protected $entityName = 'shipping_method';
+}